move isAuthenticated to /server/middleware folder
