+++ server/middleware/isAuthenticated.ts
